[Decomp][Player] Cohesion split: extract movement methods to PlayerMovement.cpp
