+++ src/game/Object/Player.cpp
@@ -3691,132 +3691,11 @@ void Player::DeleteOldCharacters(uint32 keepDays)
     sLog.outString();
 }
 
-/**
- * @brief Forces or clears rooted movement for the player.
- *
- * @param enable True to root the player; false to unroot them.
- */
-void Player::SetRoot(bool enable)
-{
-    WorldPacket data(enable ? SMSG_FORCE_MOVE_ROOT : SMSG_FORCE_MOVE_UNROOT, GetPackGUID().size() + 4);
-    data << GetPackGUID();
-    data << uint32(0);
-    SendMessageToSet(&data, true);
-}
 
-/**
- * @brief Enables or disables water walking for the player.
- *
- * @param enable True to enable water walking; false to restore normal movement.
- */
-void Player::SetWaterWalk(bool enable)
-{
-    WorldPacket data(enable ? SMSG_MOVE_WATER_WALK : SMSG_MOVE_LAND_WALK, GetPackGUID().size() + 4);
-    data << GetPackGUID();
-    data << uint32(0);
-    GetSession()->SendPacket(&data);
-}
 
-/**
- * @brief Placeholder for levitation support on this client version.
- *
- * @param enable Unused levitation state flag.
- */
-void Player::SetLevitate(bool enable)
-{
-    // TODO: check if there is something similar for 2.4.3.
-    // WorldPacket data;
-    // if (enable)
-    //    data.Initialize(SMSG_MOVE_GRAVITY_DISABLE, 12);
-    // else
-    //    data.Initialize(SMSG_MOVE_GRAVITY_ENABLE, 12);
-    //
-    // data << GetPackGUID();
-    // data << uint32(0);                                      // unk
-    // SendMessageToSet(&data, true);
 
-    // data.Initialize(MSG_MOVE_GRAVITY_CHNG, 64);
-    // data << GetPackGUID();
-    // m_movementInfo.Write(data);
-    // SendMessageToSet(&data, false);
-}
 
-/**
- * @brief Enables or disables flying movement flags for the player.
- *
- * @param enable True to enable flight-related movement flags; false to clear them.
- */
-void Player::SetCanFly(bool enable)
-{
-    WorldPacket data;
-    if (enable)
-    {
-        data.Initialize(SMSG_MOVE_SET_CAN_FLY, 12);
-    }
-    else
-    {
-        data.Initialize(SMSG_MOVE_UNSET_CAN_FLY, 12);
-    }
 
-    data << GetPackGUID();
-    data << uint32(0);                                      // unk
-    SendMessageToSet(&data, true);
-
-    data.Initialize(MSG_MOVE_UPDATE_CAN_FLY, 64);
-    data << GetPackGUID();
-    m_movementInfo.Write(data);
-    SendMessageToSet(&data, false);
-}
-
-/**
- * @brief Enables or disables feather fall movement for the player.
- *
- * @param enable True to enable feather fall; false to restore normal falling.
- */
-void Player::SetFeatherFall(bool enable)
-{
-    WorldPacket data;
-    if (enable)
-    {
-        data.Initialize(SMSG_MOVE_FEATHER_FALL, 8 + 4);
-    }
-    else
-    {
-        data.Initialize(SMSG_MOVE_NORMAL_FALL, 8 + 4);
-    }
-
-    data << GetPackGUID();
-    data << uint32(0);
-    SendMessageToSet(&data, true);
-
-    // start fall from current height
-    if (!enable)
-    {
-        SetFallInformation(0, GetPositionZ());
-    }
-}
-
-/**
- * @brief Enables or disables hover movement for the player.
- *
- * @param enable True to enable hovering; false to disable it.
- */
-void Player::SetHover(bool enable)
-{
-    WorldPacket data;
-    if (enable)
-    {
-        data.Initialize(SMSG_MOVE_SET_HOVER, 8 + 4);
-    }
-    else
-    {
-        data.Initialize(SMSG_MOVE_UNSET_HOVER, 8 + 4);
-    }
-
-    data << GetPackGUID();
-    data << uint32(0);
-    SendMessageToSet(&data, true);
-}
